Extract SourceBrowser component
