+++ app/src/DesignSystem/Components/index.js
@@ -1,9 +1,4 @@
 import Avatar from "./Avatar.svelte";
-import CommitTeaser from "./CommitTeaser.svelte";
-import File from "./File.svelte";
-import FileList from "./FileList.svelte";
-import FileSource from "./FileSource.svelte";
-import Folder from "./Folder.svelte";
 import GuideSection from "./GuideSection.svelte";
 import Notification from "./Notification.svelte";
 import NotificationFaucet from "./NotificationFaucet.svelte";
@@ -13,9 +8,9 @@ import ProjectCard from "./ProjectCard.svelte";
 import ProjectList from "./ProjectList.svelte";
 import ProjectTopbar from "./ProjectTopbar.svelte";
 import Rad from "./Rad.svelte";
-import RevisionSelector from "./RevisionSelector.svelte";
 import Sidebar from "./Sidebar.svelte";
 import SidebarLayout from "./SidebarLayout.svelte";
+import SourceBrowser from "./SourceBrowser.svelte";
 import StepCounter from "./StepCounter.svelte";
 import Swatch from "./Swatch.svelte";
 import TxRow from "./TxRow.svelte";
@@ -23,11 +18,6 @@ import UserCard from "./UserCard.svelte";
 
 export {
   Avatar,
-  CommitTeaser,
-  File,
-  FileList,
-  FileSource,
-  Folder,
   GuideSection,
   Notification,
   NotificationFaucet,
@@ -37,9 +27,9 @@ export {
   ProjectList,
   ProjectTopbar,
   Rad,
-  RevisionSelector,
   Sidebar,
   SidebarLayout,
+  SourceBrowser,
   StepCounter,
   Swatch,
   TxRow,
